new version fixing almost all problems

+++ interpdata/interpdata.h
@@ -8,8 +8,10 @@ class Interpdata
     // data members
     std::string d_subname;
     size_t d_nTrial;
-    size_t d_beginInterp;
-    size_t d_endInterp;	
+//     size_t d_beginInterp;
+//     size_t d_endInterp;	
+    double d_beginInterp;
+    double d_endInterp;	
     double d_begX;
     double d_endX;
     double d_begY;
@@ -23,8 +25,10 @@ class Interpdata
 	
 	void setSub(std::string sub);
 	void setNtrial(size_t ntrial);
-	void setBegin(size_t beginInterp);
-	void setEnd(size_t endInterp);
+// 	void setBegin(size_t beginInterp);
+// 	void setEnd(size_t endInterp);
+	void setBegin(double beginInterp);
+	void setEnd(double endInterp);
 	void setBegX(double begx);
 	void setEndX(double endx);
 	void setBegY(double begy);
@@ -35,8 +39,10 @@ class Interpdata
 
 	std::string sub() const;
 	size_t nTrial() const;
-	size_t iBegin() const;
-	size_t iEnd() const;
+// 	size_t iBegin() const;
+// 	size_t iEnd() const;
+	double iBegin() const;
+	double iEnd() const;
 	double begX() const;
 	double endX() const;
 	double begY() const;
