+++ interpdata/setBeginInterp.cc
@@ -1,6 +1,7 @@
 #include "interpdata.ih"
 
-void Interpdata::setBegin(size_t beginInterp)
+//void Interpdata::setBegin(size_t beginInterp)
+void Interpdata::setBegin(double beginInterp)
 {
   d_beginInterp = beginInterp;
 }
