+++ interpdata/setEndInterp.cc
@@ -1,6 +1,7 @@
 #include "interpdata.ih"
 
-void Interpdata::setEnd(size_t endInterp)
+//void Interpdata::setEnd(size_t endInterp)
+void Interpdata::setEnd(double endInterp)
 {
   d_endInterp = endInterp;
 }
